#167 Refactor: rename method, rename files
