+++ javascript/big-practice/src/scripts/main.js
@@ -1,26 +1,23 @@
 import ListNoteView from './views/listNoteView';
 import NoteModel from './models/noteModel';
 import NoteController from './controllers/noteController';
-import AuthenticationController from './controllers/authenticationController';
-import AuthenticationModel from './models/authenticationModel';
+import UserController from './controllers/userController';
 import LoginView from './views/loginView';
 
 const listNoteModel = new NoteModel();
 
 const listNoteView = new ListNoteView();
 
-const authenticationModel = new AuthenticationModel();
-
 const loginView = new LoginView();
 
 const listNoteController = new NoteController(listNoteModel, listNoteView);
-const authenticationController = new AuthenticationController(authenticationModel, loginView);
+const userController = new UserController(loginView);
 
 (() => {
   const page = document.body.className;
 
   if (page === 'login-page') {
-    authenticationController.init();
+    userController.init();
   } else {
     listNoteController.init();
   }
