+++ javascript/big-practice/src/scripts/models/userModel.js
@@ -8,6 +8,32 @@ export default class UserModel {
   }
 
   /**
+   * @description function get user by email. Check email is available or not
+   * and if user is available it will check password is correct or not
+   *
+   * @param {String} email is email take from input email login
+   * @param {String} password is password take from input password login
+   *
+   * @returns {Boolean, Boolean} isEmail, isPassword
+   */
+  async verifyCredential(email, password) {
+    const users = await getUser(email, 'email');
+    let isEmail = false;
+    let isPassword = false;
+
+    if (users.length) {
+      if (users[0].password === password) {
+        this.localStorage.setItems(STORAGE_KEYS.USER_ID, users[0].id);
+        isPassword = true;
+      }
+
+      isEmail = true;
+    }
+
+    return { isEmail, isPassword };
+  }
+
+  /**
    * @description find user by id
    *
    * @returns {String} email or Unknown
